Packaged all Classes into a single package named duke
