+++ src/main/java/duke/Parser.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.time.format.DateTimeParseException;
 
 public class Parser {
